Move Discord webhook logic into outgoing-webhooks
