+++ apps/website/src/server/notifications.ts
@@ -16,7 +16,7 @@ import {
 
 import { prisma } from "@/server/db/client";
 import { callEndpoint } from "@/server/utils/queue";
-import { triggerDiscordChannelWebhook } from "@/server/discord";
+import { triggerDiscordChannelWebhook } from "@/server/outgoing-webhooks";
 import { escapeLinksForDiscord } from "@/utils/escape-links-for-discord";
 
 import type { CreatePushesOptions } from "@/pages/api/notifications/batched-create-notification-pushes";
