+++ apps/website/src/server/outgoing-webhooks.ts
@@ -1,6 +1,7 @@
 import type { OutgoingWebhook } from "@prisma/client";
 
 import { prisma } from "@/server/db/client";
+import { env } from "@/env";
 
 export type OutgoingWebhookType = "form-entry" | "unknown";
 
@@ -147,4 +148,94 @@ export async function retryOutgoingWebhooks({
   );
 
   await Promise.all(tasks);
+}
+
+export const OUTGOING_WEBHOOK_TYPE_DISCORD_CHANNEL = "discordChannel";
+
+type DiscordWebhookNotificationBody = {
+  avatar_url: string;
+  username: string;
+  content?: string;
+  tts: boolean;
+  allowed_mentions: {
+    parse?: Array<"everyone" | "roles" | "users">;
+    users?: string[];
+  };
+  embeds?: Array<{
+    title: string;
+    description: string;
+    color?: number;
+    fields?: Array<{ name: string; value: string; inline?: boolean }>;
+    footer?: { text: string; icon_url?: string };
+    thumbnail?: { url: string };
+    image?: { url: string };
+    author?: { name: string; url?: string; icon_url?: string };
+    timestamp?: Date;
+  }>;
+};
+
+export async function triggerDiscordChannelWebhook({
+  webhookUrl,
+  contentTitle,
+  contentMessage,
+  contentLink,
+  imageUrl,
+  botName: username = env.DISCORD_CHANNEL_WEBHOOK_NAME,
+  expiresAt,
+  toEveryone = false,
+}: {
+  webhookUrl: string;
+  contentTitle?: string;
+  contentMessage?: string;
+  contentLink?: string;
+  imageUrl?: string;
+  botName?: string;
+  expiresAt?: Date;
+  toEveryone?: boolean;
+}) {
+  const embed = {
+    title: contentTitle || "Notification",
+    description: (
+      (contentMessage || "") +
+      (contentLink
+        ? `\n[${contentLink.replace(/^https?:\/\/(www\.)?/, "")}](${contentLink})`
+        : "")
+    ).trim(),
+    color: 0x636a60,
+    url: contentLink,
+    footer: {
+      text: "alveus.gg/updates",
+      icon_url: `${env.NEXT_PUBLIC_BASE_URL}/apple-touch-icon.png`,
+    },
+    timestamp: new Date(Date.now()),
+    image: {
+      url: imageUrl || "",
+    },
+  };
+
+  const body: DiscordWebhookNotificationBody = {
+    username,
+    tts: false,
+    avatar_url: `${env.NEXT_PUBLIC_BASE_URL}/apple-touch-icon.png`,
+    allowed_mentions: {
+      parse: [],
+    },
+    embeds: [embed],
+  };
+
+  if (toEveryone) {
+    body.allowed_mentions = { parse: ["everyone"] };
+    body.content = `@everyone ${contentTitle}`;
+  }
+
+  const url = new URL(webhookUrl);
+  url.searchParams.set("wait", "true");
+
+  return triggerOutgoingWebhook({
+    url: url.toString(),
+    type: OUTGOING_WEBHOOK_TYPE_DISCORD_CHANNEL,
+    body: JSON.stringify(body),
+    retry: true,
+    expiresAt,
+  });
 }
